rename: 재사용을 위한 도메인명 수정
