+++ src/entities/portfolio-details/model/adapters.ts
